Renamed to Async to be more in line with other AsyncXXX classes
