+++ docs/execution.rst
@@ -346,7 +346,7 @@ The graphql-java engine ensures that all the ``CompletableFuture`` objects are c
 that follows the graphql specification.
 
 There is a helpful shortcut in graphql-java to create asynchronous data fetchers. 
-Use ``graphql.schema.AsynchronousDataFetcher.async(DataFetcher<T>)`` to wrap a 
+Use ``graphql.schema.AsyncDataFetcher.async(DataFetcher<T>)`` to wrap a
 ``DataFetcher``. This can be used with static imports to produce more readable code.
 
 .. code-block:: java
